MODIFIED: Pass variable binding term through the pengine itself.  As a result, the success event changed from `success(Pengine, Answer, Time, More)` to `success(Pengine, Answer, Projection, Time, More)` and the VarNames argument has been removed from pengines:write_result/4 and event_to_json/4 (both now accepting 3 arguments).
This largely simplifies the design.  In addition, it provides Pengines
access to the actual variable names inside a query and the possibility
to control the projection variables (in the future).

Code that defines additional output formats, such as writing CSV files,
must be changed.  Most typical usage of Pengines should not be affected.