ADDED: Allow no-chunking behavior for pengines.
If `chunk=false` we avoids the usage of
findnsols (and the effects on global variables associated with its
backtracking). Initially, this was motivated by the fact that I wanted
the pengine top-level (which is recursive in nature) to respect the chr
store, but I also think it's a more principled thing to do.