Allow no-chunking behavior for pengines.
The old options chunk(integer) is now optional. In the case it's not
set, we fall back to a non-chunking behavior that avoids the usage of
findnsols (and the effects on global variables associated with its
backtracking). Initially, this was motivated by the fact that I wanted
the pengine top-level (which is recursive in nature) to respect the chr
store, but I also think it's a more principled thing to do.